♻️ Move hasUntrackedFiles from guard.got o diff.go
