+++ diff.go
@@ -12,3 +12,12 @@ func hasUncommittedChanges() (bool, error) {
 	}
 	return len(bytes.TrimSpace(output)) > 0, nil
 }
+
+func hasUntrackedFiles() (bool, error) {
+	cmd := exec.Command("git", "ls-files", "--others", "--exclude-standard")
+	output, err := cmd.Output()
+	if err != nil {
+		return false, err
+	}
+	return len(bytes.TrimSpace(output)) > 0, nil
+}
